+++ extensions/extensions.go
@@ -1,0 +1,5 @@
+package extensions
+
+import "github.com/knackwurstking/tgs/pkg/extension"
+
+var register = []extension.Extension{} // Containing all the extensions registered