+++ pkg/extension/extension.go
@@ -1,0 +1,4 @@
+package extension
+
+// NOTE: `any` is just a placeholder
+type Extension any